requirements.txt
pandas 
faker
apache-airflow-providers-apache-kafka
confluent-kafka
kafka-python
